Double the fun

+++ eLime.NetDaemonApps.Config/SmartIrrigation/AntiFrostMistingIrrigationConfig.cs
@@ -3,8 +3,8 @@ namespace eLime.NetDaemonApps.Config.SmartIrrigation;
 public class AntiFrostMistingIrrigationConfig
 {
     public string TemperatureEntity { get; set; }
-    public int CriticalTemperature { get; set; }
-    public int LowTemperature { get; set; }
+    public double CriticalTemperature { get; set; }
+    public double LowTemperature { get; set; }
     public TimeSpan MistingDuration { get; set; }
     public TimeSpan MistingTimeout { get; set; }
 }
